genomics: use ephemeral memory to avoid fragmentation
Higher memory usage occurs when Spate is compiled with the Cray compiler
or with the IBM XL compiler, but does not occur when compiled with the GNU
compiler. This indicates that somehow a different memory engine was used
to fulfill the service provided with malloc()/free() with these toolchains.

I figured out that this was memory fragmentation after our collaborator
Carlos Sosa (Cray Inc.) pointed out that the Cray toolchain linking option
'-hsystem_alloc' usually solves this problem, but otherwise still indicates
that the code does a lot of small allocations.

0xc170626e is the key for MEMORY_MEMORY_POOL (anonymous pool)

Before:
 376058 0xc170626e

After:
 147266 0xc170626e

(-39.16%)

Reported-by: Huy Bui <huybui@anl.gov>
Link: https://github.com/GeneAssembly/biosal/issues/722
Link: https://github.com/GeneAssembly/biosal/issues/729
Cc: Carlos P. Sosa <cpsosa@cray.com>
Signed-off-by: Sébastien Boisvert <boisvert@anl.gov>