genomics: use memory pool for visited DNA sequences
Memory fragmentation exhausts the usable memory in the heap
of the virtual memory of the process.

By using the memory pool for visited vertices, the number of calls
to core_memory_allocate() is lowered, thus reducing memory fragmentation.

Below are the numbers of calls to core_memory_allocate with key= 0xc170626e.
This key corresponds to the "anonymous" memory pool.

Before: 123918 0xc170626e
After: 58219 0xc170626e
Change: -53.01%

Link: https://github.com/GeneAssembly/biosal/issues/729
Link: https://github.com/GeneAssembly/biosal/issues/684
Signed-off-by: Sébastien Boisvert <boisvert@anl.gov>