thorium: use the multiplexer only on specific messages
Multiplexing every message with a buffer size below the threshold
slows down things because such messages will only leave the node
when the timeout is reached (or in the next call if the timeout is 0).

This patch instead requires that actions to be multiplexed
be explicitly registered in the multiplexer policy.

Link: https://github.com/GeneAssembly/biosal/issues/749
Signed-off-by: Sébastien Boisvert <boisvert@anl.gov>