thorium: refine the message deleguation model
Before, the worker pool would push a message in the mailbox
of an actor and then notify the worker of that event.

Now, the worker pool simply push the message to the worker,
and the worker is responsible for giving the message to the actor
and also scheduling the actor.

Signed-off-by: Sébastien Boisvert <boisvert@anl.gov>