thorium: improve the code that manages message metadata
Additional information is added to any message in order to track things such
as source actor, destination actor, and action specifier. The code that increase
the value for the number of bytes in the buffer of a message was confusing.

This patch improves how the metadata is managed: typically, the metadata is
added by workers only and is removed by workers only. The node layer does not
need to know about the content of the message (except for the dispatch code path).

Signed-off-by: Sébastien Boisvert <boisvert@anl.gov>