thorium: don't use rand() to avoid system calls (sys_futex)
The function rand() in the glibc (used in most GNU/Linux distributions)
generates 2 system calls on Linux (sys_futex with FUTEX_WAIT or FUTEX_WAKE).
Therefore, this function can not be used in Thorium.

The links below document where the locks are located in the rand() code path
when using the GNU glibc. I don't know where __libc_lock_lock() is defined
though.

Link: https://sourceware.org/git/?p=glibc.git;a=blob;f=stdlib/rand.c#l27
Link: https://sourceware.org/git/?p=glibc.git;a=blob;f=stdlib/random.c#l292
Signed-off-by: Sébastien Boisvert <boisvert@anl.gov>